--- src/dataSnatcher.js
@@ -1,21 +1,0 @@
-/* dataSnatcher.js - snatches data from the business API
- */
-const express = require('express');
-const DataMaster = require('./DataMaster');
-const auraList = require('../sample-data/aura/auras.json');
-
-const router = express.Router();
-const dataMaster = new DataMaster();
-
-// router routes to MongoDB
-
-router.get('/api/businesses', (req, res) => {
-  dataMaster.find(req, res);
-});
-
-// GET request for aura list
-router.get('/auras', (req, res) => {
-  res.json(auraList);
-});
-
-module.exports = router;
